call old JUnit4 StringConverter#convert if set instead of always new #convert
With reimplementing `DataConverter` in core/ the old JUnit4
`StringConverter#convert(...)` was not called any more. Instead always
the new `StringConverter#convert(...)` of core/ was called. This is
happend by accident because old one extends new one such that new API
was available, too. Now, this problem is fixed. It occured since v2.0.

Signed-off-by: Andreas Schmid <service@aaschmid.de>